requirements.txt
bleak~=0.20.2
pillow~=10